Properties Lock

+++ SusSnakeClient/BodyClass.cs
@@ -3,8 +3,8 @@ using System.Text.Json.Serialization;
 public class BodyClass
 {
     [JsonPropertyName("x"), JsonInclude]
-    public float X { get; set; }
+    public int X { get; set; }
 
     [JsonPropertyName("y"), JsonInclude]
-    public float Y { get; set; }
+    public int Y { get; set; }
 }
