+++ SusSnakeClient/SnakeProperties.cs
@@ -4,10 +4,10 @@ using Raylib_cs;
 public class SnakeProperties
 {
     [JsonPropertyName("x"), JsonInclude]
-    public float X { get; set; }
+    public int X { get; set; }
 
     [JsonPropertyName("y"), JsonInclude]
-    public float Y { get; set; }
+    public int Y { get; set; }
 
     [JsonPropertyName("body"), JsonInclude]
     public List<BodyClass> Body { get; set; } = new();
